requirements.txt
flask==2.0.1
requests==2.26.0
python-dotenv==0.19.0
anthropic==0.18.0
pytube==15.0.0
youtube-transcript-api==0.6.1
gunicorn==20.1.0